requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0

# Numerical computing
numpy>=1.21.0

# Utilities
natsort>=8.0.0
tqdm>=4.65.0
